refactor: Add IP address to device connection logging
This commit modifies the addConnection function in connectionManager.js to include the IP address of the connected device in the log message. It also updates the corresponding log message in websocketHandler.js to include the IP address. These changes provide more detailed information about the connected devices in the logs, improving troubleshooting and monitoring capabilities.